+++ src/MacroTools/BookSystem/Powers/PowerCard.cs
@@ -1,7 +1,8 @@
 ﻿using MacroTools.FactionSystem;
+using MacroTools.Frames;
 using static War3Api.Common;
 
-namespace MacroTools.Frames.Books.Powers
+namespace MacroTools.BookSystem.Powers
 {
   public sealed class PowerCard : Card
   {
